Address Copilot review: fix Lock _owningThreadId type and ComWrappers null handling
- Lock.cs: read _owningThreadId as int (matches CoreLib declaration)
  so TargetFieldExtensions.AssertPrimitiveType debug validation passes;
  cast to uint at the SyncBlock_1 API boundary where the public
  TryGetLockInfo signature returns uint.
- ComWrappers.cs: use TryGetStaticFieldAddress and return
  TargetPointer.Null when the ComWrappers statics base isn't allocated,
  so callers (GetMOWs / GetComWrappersRCWForObject) report 'no data'
  via the existing null-check path instead of throwing.

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/Contracts/SyncBlock_1.cs
@@ -51,7 +51,7 @@ internal readonly struct SyncBlock_1 : ISyncBlock
             bool monitorHeld = (lockData.State & 1) != 0;
             if (monitorHeld)
             {
-                owningThreadId = lockData.OwningThreadId;
+                owningThreadId = (uint)lockData.OwningThreadId;
                 recursion = lockData.RecursionCount;
             }
             return monitorHeld;
